+++ src/qa/menu/ecckey_menu.cpp
@@ -8,7 +8,7 @@
 
 namespace ns_menu
 {
-	int main_menu::fECCKey(int choice)
+	int main_menu::fECCKey(size_t choice)
    	{
 		int r = 0;
